+++ packages/recorder/src/index.ts
@@ -1,0 +1,12 @@
+export { hashRequest } from "./hash.js";
+export { scrub, detectUnscrubbed, type ScrubConfig } from "./scrub.js";
+export {
+  createRecordingFetch,
+  type RecorderOptions,
+  type RecorderMode,
+} from "./fetch-recorder.js";
+export {
+  findStaleFixtures,
+  type StalenessCheckOptions,
+  type StaleFixture,
+} from "./staleness.js";